+++ examples/typescript/src/gen/models.ts
@@ -243,7 +243,6 @@ export type FullAddress = Address & {
    * @type string
    */
   streetNumber: string
-} & {
   /**
    * @type string
    */