+++ examples/typescript/src/gen/modelsLiteral.ts
@@ -231,7 +231,6 @@ export type FullAddress = Address & {
    * @type string
    */
   streetNumber: string
-} & {
   /**
    * @type string
    */