+++ packages/plugin-ts/src/generators/__snapshots__/createPetWithUnknownTypeVoid.ts
@@ -11,6 +11,10 @@ export type CreatePets201 = error & {
    * @type object | undefined
    */
   name?: errorCode
+  /**
+   * @type string
+   */
+  message: string
 }
 
 /**